requirements.txt
Django==1.8
djangorestframework==3.6.3
funcsigs==1.0.2
mock==2.0.0
pbr==3.1.1
pytz==2017.2
six==1.10.0
